backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pyjwt
bcrypt
pymongo
aiosmtplib
dotenv
